+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2022-10-13 - Bugfix: Custom SCSS and background image SCSS was included twice, solves #103 #104.
 * 2022-10-12 - Settings: screenshot image for theme chooser solves #33.
 
 ### v4.0-r3